Font rendering fix for < IE9
Fabric expects text to be rendered as Canvas elements. Cufon however falls back to VML rendering when canvas is not available in the browser. This mod of the Cufon library ALWAYS enabled the canvas rendering engine for Cufon.  This way font-rendering is available on IE7 and IE8 when excanvas is available.