Changed getPointer() to look for any scrolled descendants. Also checks for fixed positioned elements.
I also changed canvas.containsPoint() to take in an optional 'cached'
pointer object. Should save a bit of time when checking a large number
of objects, especially with the added work in getPointer().
getPointer() no longer calls pointerX() or pointerY(), so I don't think
they're needed. Didn't want to get rid of them since there were a few
comments in pointerX() them that seemed relevant, but I'm not sure how
to test for that.