Smooth free drawing
While drawing on contextTop canvas we show a smooth curved line using
'quadraticCurve' canvas methode.

On mouse up we convert the collected mousepoints `_freeDrawingPoints`to a path
fabric object using quadtratic bezier curveto syntax.

Note:

If we do a lot of 'live drawing while mouse down' and add the shape on mouse up
we may try to refactor and create a new Live Render canvas class ?

So far only `pen` has this behavior but we could want to draw rectangles or
straight lines the same way.