Fix fontStyle 'italic' for non system (@font-face)
Most @font-face fonts do not support the native 'italic' declaration when rendered on a canvas.  This fix uses a small transform instead.  This could be further improved by checking if the 'fontFamily' is a non system font and only falling back to the transform method when needed.  To my eyes, however, this method looks exactly the same as the native one.