Add text:selected event
Add event that fires whenever text selection is changed.  This is extremely useful when developing toolbars to manipulate the text styles. e.g. when selection changes an app can then check the characters in the selection to see if they are bold.  If so a bold button is highlighted etc.