Misc changes to viewbox handles
I'm not so in the topic to countinue the changes.
What i wanted to point out is the use of viewbox first 2 params, normally 0 and 0, but when those are negatives the mess start. This solves even some bounding box problems.
I edited just circles and paths , cause i have no other examples of failing svg.