Update parser.js
if we delete it , it won't pass to child graphic elements.
if this fillopacity is specified on a container element, it will change color of container element and then disappear. Color would then go to child element.
But if fill is specified on child element and fillopacity on parent element, this has to stay.