Updated fabric.js, added mouseout on upperCanvas element
There are **mouseover** & **mouseout** events for objects, which works fine. But there is a case where **mouseout** on an object does not trigger. Lets assume we have an overlay div on the main canvas and that div is placed right below the object. Now when you hover over the object **mouseover** gets triggered. But when you move your mouse to the overlay div the **mouseover** does not trigger. That is because these events are bind to **upper-canvas** and when mouse is on the overlay div the **mouseover** and **mouseout** does not gets trigger. 

Or lets forget about overlay div. Lets assume the object is on the left. And when you mouse out of the object to left in this [fiddle](https://jsfiddle.net/45vfbgw1/2/) you will notice the **mouseout** does not trigger. Because now mouse is out of the main canvas.

So i have added a **mouseout** event for **upper-canvas**. When this event gets trigger it will call the **mouseout** of the last **mouseover** object.

I have created 2 fiddles to explain the issue.

[Fiddle using the current fabirc.js:](https://jsfiddle.net/45vfbgw1/) In this fiddle you can see when you **mouseover** the big red box the **mouseover** event triggers. But when you move the mouse right below the object to the overlay div, the **mouseout** does not trigger.

[Fiddle with solution:](http://jsfiddle.net/wm69g3ed/)