[#94042944] Fix uniform scaling with lockScalingFlip.
Previously when enabling `lockScalingFlip` with uniform scaling,
the scale would be denied prematurely if the local x or y mouse position
became negative. However, this is often well before the effective scale
would actually result in a flip, so results in awkward and surprising behaviour.

Now the decision of whether to deny scaling based on `lockScalingFlip` is
delegated to `_scaleObjectEqually` along with the rest of the operation.