[#94464656] Ensure that the text wrapping algorithm terminates
Checking that the max width is narrower than the biggest letter should
only be an optimisation, not a guarantee relied on to not lock up the
entire sketchpad!

In any case, 'm' is bigger than 'W' in the font we use.

I have adjusted the text wrapping algorithm to detect that a single letter
is not going to fit on the line, and just make the line consist of that
letter & move on. This should rarely have to run, because the < largest
letter optimisation should catch most of it.