[#97068140] Don't wrap inside words.
Instead, if a word can't even fit on its own line, just let it
overflow beyond the horizontal bounds of the shape.

This has the dual benefits of not mangling your text (really, wh
en are you going to be pleased by having your word broken in the
middle?) and entirely avoids the crazy ~exponential-time algorithm
that implemented that behaviour.

The main downside of this is that the appearance of text while
editing is now more often different from how it will render once
you stop editing. The few people that were involved in playing
with this change (KJ, DRP, JP, RM-ish) were reasonably convinced
that this is an acceptable compromise.