[#97632112] Stop constraining the scale of objects
This is actually a real problem. The stroke width calculation is all
wrong when the object scale gets constrained; it ends up continuing to
grow the stroke width but not the object size, rapidly causing
everything to be covered in stroke. This feature is doing more harm than
good, so I'm just getting rid of it.