Fix to `animateColor` when transitioning to/from opacity of 0
Changes:
- The red, green, and blue RGBA color parameters are now rounded
using `Math.round` rather than truncated via `parseInt` to ensure
that a value of 0 can be attained.
- The alpha parameter of the calculated RGBA colors within
animateColor will now round to the nearest thousandth. This prevents
a near-zero number represented in scientific notation from being
used as an RGBA color's alpha parameter (which is undesirably the
render-equivalent of an alpha parameter of 1).