Add changes to 'What To Use Fabric.js For..' sect.
I added the "What To Use Fabric.js For And When To Use It" heading for the section.  I added a bit more description for when using Fabric.js with custom t-shirts.  I also added the feature of converting SVG to canvas objects and vice verse to the section.  I removed the last two lines of the paragraph because they sounded like part of the introduction paragraph in this file and my non-technical section that I added on my branch too.