[#128922461] Fix group rotation of lines
Our stileSnappedScaling behaviour does its maths in canvas space and updates (x1,y1,x2,y2) with canvas space coordinates. This means we have to remove any other transformation on the object, otherwise they will be applied to the new coordinates on every change (causing the "flying baton" bug).

We were previously removing the rotation in the setter for the Line, but this meant that any Lines that were actually rotated (only achievable if rotated in a group - a Line can't be rotated by itself) had their rotation stripped during deserialization.