Ensure clone invokes callback in all cases
> Clones an instance, some objects are async, so using callback method will work for every object. 

This should mean that `callback` is invoked no matter what, however if `this.constructor.fromObject` is not a function then it should never be invoked. 

This solves the issue by invoking `callback` if the value is just being returned. 

`setTimeout` is used to ensure `callback` is invoked in an *async* fashion