Fix bug - onMouseUp
If the mouse is held down and dragged off the "corner", then `__onMouseUp` **can not find the corner** because the pointer is no longer positioned over the corner. As a result, the `mouseUpHandler` is not invoked when the mouse is released.

This fixes that bug, by using the `__corner` assigned during `__onMouseDown`.