Update scale.ts
When scaling from any one of tl, tr, br, bl, ml, mt, mr or mb controls and you are not 100% centered on it you will get a jump effect. This commit helps to fix that by adjusting the starting position by the offset amount when dragging a scale control.

This is most noticeable when using larger controls. Works in both ALT and non ALT modes. On Desktop and Mobile.

Rotate and Move are not affected.