The selection and cursor rendering of IText objects are now clipped to the object's clip path and its ancestors' clip paths.
This change is WIP in attempt to address [the issue](https://github.com/fabricjs/fabric.js/issues/10274) where the selection and cursor were not being clipped correctly when the object was clipped.

A `applyClipPathToCtx` method has been added to apply the clip paths to the rendering context. A `findClipPathAncestors` method has been added to find the clip path ancestors of the object.

The `renderCursorOrSelection` method has been updated to use the `applyClipPathToCtx` and `findClipPathAncestors` methods to clip the selection and cursor rendering.

Currently only partially works — the selection is  positioned and clipped but the rect is black.